moved all motion code into a separate folder and organised them. Also made file locations relative to folder location
