+++ GameFiles/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\Ghost\\AppData\\Local\\Programs\\Python\\Python36\\python.exe"
+}
