+++ GameFiles/cards.txt
@@ -1,0 +1,6 @@
+Writing this to a file.
+BLACK 5, 7, 1
+RED 5, 3, 5
+Black 7, 3, 2
+RED 9, 6, 8
+BLACK 2, 6, 2
