+++ GameFiles/pick.txt
@@ -1,0 +1,1 @@
+RED 5, 3, 5
