+++ README.md
@@ -1,2 +1,13 @@
 # TSE_Project
 Using QiBullet to test our pepper code
+
+The GameLogic file is the main program that calls the other sections of code:
+SpawnPepperTestPointing: which starts the virtual simulation and enables pepper to move
+CardScanner: reads in information from the text file
+
+you will need to use pip install to install any imported modules you may need such as numpy or the q
+
+needed modules:
+qibullet
+cv2
+numpy
