--- pick.txt
@@ -1,1 +1,0 @@
-RED 5, 3, 5
